requirements.txt
crewai
openai
python-telegram-bot

